Fix missing theme reference and a some more bindings
Don't use the default paste buffer for single character deletes. Also
use ^C to exit modes since it does what I was trying to do before. This
allows me to use Ctrl with the hjkl in insert mode to move around. It
does remove at two existing mapping of ^H for back delete and ^K for
inserting a representation of the next typed key.